@@ -8,14 +8,18 @@
   import { Label } from '$lib/components/ui/label';
   import { Skeleton } from '$lib/components/ui/skeleton';
 
-  export let modalVisible: boolean;
-  export let onSaved: () => void;
-  let showArchive = $auth.githubSettings.archive;
-  let state = $auth.githubSettings.state;
-  let type = $auth.githubSettings.type;
-  let selectedOrganizations = $auth.githubSettings.organizations || [];
-  let organizationOptions: Array<{ value: string; label: string }> = [];
-  let loadingOrganizations = false;
+  interface Props {
+    modalVisible: boolean;
+    onSaved: () => void;
+  }
+
+  let { modalVisible = $bindable(), onSaved }: Props = $props();
+  let showArchive = $state($auth.githubSettings.archive);
+  let ghState = $state<string>($auth.githubSettings.state || 'open');
+  let ghType = $state<string>($auth.githubSettings.type || 'review-requested');
+  let selectedOrganizations = $state<string[]>($auth.githubSettings.organizations || []);
+  let organizationOptions: Array<{ value: string; label: string }> = $state([]);
+  let loadingOrganizations = $state(false);
 
   const changeShowArchive = (checked: boolean) => {
     showArchive = checked;
@@ -24,21 +28,13 @@
   const onSave = () => {
     $auth.updateGithubSettings({
       archive: showArchive,
-      state,
-      type,
+      state: ghState as typeof $auth.githubSettings.state,
+      type: ghType as typeof $auth.githubSettings.type,
       organizations: selectedOrganizations,
     });
     modalVisible = false;
     onSaved();
   };
-
-  const updateSelectedOrganizations = (target: any) => {
-    const incoming = target as { value: string; label: string }[];
-    selectedOrganizations = incoming.map(i => i.value);
-  };
-
-  const updateType = (target: any) => (type = target.value);
-  const updateState = (target: any) => (state = target.value);
 
   const typeOptions = [
     { value: 'review-requested', label: 'Reviews' },
@@ -51,6 +47,16 @@
     { value: 'closed', label: 'Closed' },
     { value: 'all', label: 'All' },
   ];
+
+  const getTriggerContent = (value: string[] | string, placeholder?: string, multipleMessage?: string) => {
+    if (Array.isArray(value)) {
+      if (value.length === 0) return placeholder;
+      if (value.length === 1) return value[0];
+      return value.length + ' ' + multipleMessage;
+    } else {
+      return [...typeOptions, ...stateOptions].find(option => option.value === value)?.label || placeholder;
+    }
+  };
 
   onMount(() => {
     if (!$auth.account) return;
@@ -72,9 +78,9 @@
       <Label for="archive">Show archived</Label>
     </div>
     <div class="flex items-center justify-between gap-4">
-      <Select.Root items={typeOptions} selected={typeOptions.find(v => v.value === type)} onSelectedChange={updateType}>
+      <Select.Root type="single" items={typeOptions} onValueChange={v => (ghType = v)}>
         <Select.Trigger>
-          <Select.Value placeholder="Type" />
+          {getTriggerContent(ghType)}
         </Select.Trigger>
         <Select.Content>
           {#each typeOptions as { value, label }}
@@ -82,13 +88,9 @@
           {/each}
         </Select.Content>
       </Select.Root>
-      <Select.Root
-        items={stateOptions}
-        selected={stateOptions.find(v => v.value === state)}
-        onSelectedChange={updateState}
-      >
+      <Select.Root items={stateOptions} type="single" onValueChange={v => (ghState = v)}>
         <Select.Trigger>
-          <Select.Value placeholder="State" />
+          {getTriggerContent(ghState)}
         </Select.Trigger>
         <Select.Content>
           {#each stateOptions as { value, label }}
@@ -102,13 +104,13 @@
         <Skeleton class="h-10 w-full" />
       {:else if organizationOptions.length !== 0}
         <Select.Root
+          type="multiple"
           items={organizationOptions}
-          selected={selectedOrganizations.map(v => ({ value: v, label: v }))}
-          multiple
-          onSelectedChange={updateSelectedOrganizations}
+          bind:value={selectedOrganizations}
+          onValueChange={v => (selectedOrganizations = v)}
         >
           <Select.Trigger class="w-full">
-            <Select.Value placeholder="Organizations" />
+            {getTriggerContent(selectedOrganizations, 'Select organizations', 'organizations selected')}
           </Select.Trigger>
           <Select.Content>
             {#each organizationOptions as { value, label }}
@@ -120,6 +122,6 @@
     </div>
   </div>
   <div class="relative mt-4 flex items-end justify-end">
-    <Button type="button" size="sm" on:click={onSave}>Save</Button>
+    <Button type="button" size="sm" onclick={onSave}>Save</Button>
   </div>
 </div>
