@@ -2,7 +2,7 @@
   import { onMount } from 'svelte';
   import { getVersion, getName } from '@tauri-apps/api/app';
   import { auth, defaultSettings } from '../lib/auth';
-  import Range from './Range.svelte';
+  import { Slider } from '$lib/components/ui/slider/index.js';
   import Theme from './Theme.svelte';
   import { isEnabled } from '../lib/auto-start';
   import { Button } from '$lib/components/ui/button';
@@ -10,15 +10,16 @@
   import { Switch } from '$lib/components/ui/switch';
   import { Label } from '$lib/components/ui/label';
 
-  export let modalVisible: boolean;
-  let openAtStartup = $auth.settings?.openAtStartup;
-  let isCompactMode = $auth.settings?.isCompactMode;
-  let fetchInterval = ($auth.settings.fetchInterval || defaultSettings.fetchInterval) / 1000;
-  let app = { name: '', version: '' };
+  interface Props {
+    modalVisible: boolean;
+  }
 
-  const changeAutoStart = (checked: boolean) => {
-    openAtStartup = checked;
-  };
+  let { modalVisible = $bindable() }: Props = $props();
+  let openAtStartup = $state($auth.settings?.openAtStartup);
+  let isCompactMode = $state($auth.settings?.isCompactMode);
+  let fetchInterval = $state(($auth.settings.fetchInterval || defaultSettings.fetchInterval) / 1000);
+  let app = $state({ name: '', version: '' });
+
   const changeCompactMode = (checked: boolean) => {
     isCompactMode = checked;
   };
@@ -36,6 +37,8 @@
       const [name, version] = values;
       app = { name, version };
     });
+    // Not sure why invoke('plugin:autostart|is_enabled') returns false when local storage is set to true
+    // Application priviliges in dev perhaps?
     isEnabled().then(active => {
       openAtStartup = active;
     });
@@ -45,9 +48,14 @@
 <div class="space-y-6">
   <Theme />
   <!-- <Toggle name="open_at_start" checked={openAtStartup} label="Auto start Gitbar" on:change={changeAutoStart} /> -->
-  <div class="flex flex-row gap-2 items-center justify-between rounded-lg border p-4">
+  <div class="flex flex-row items-center justify-between gap-2 rounded-lg border p-4">
     <div class="flex items-center space-x-2">
-      <Switch id="open_at_start" name="open_at_start" checked={openAtStartup} onCheckedChange={changeAutoStart} />
+      <Switch
+        id="open_at_start"
+        name="open_at_start"
+        checked={openAtStartup}
+        onCheckedChange={c => (openAtStartup = c)}
+      />
       <Label for="open_at_start">Auto start Gitbar</Label>
     </div>
     <div class="flex items-center space-x-2">
@@ -57,19 +65,13 @@
     <!-- <Toggle name="compact_mode" checked={isCompactMode} label="Compact mode" on:change={changeCompactMode} /> -->
   </div>
   <div class="rounded-lg border p-4">
-    <label for="fetch_interval" class="block text-sm font-bold text-gray-700 dark:text-gray-100 mb-4">
+    <label for="fetch_interval" class="mb-4 block text-sm font-bold text-gray-700 dark:text-gray-100">
       Fetch interval <strong>{fetchInterval} sec</strong>
     </label>
-    <Range
-      on:change={e => (fetchInterval = e.detail.value)}
-      initialValue={fetchInterval}
-      min={5}
-      max={60}
-      id="fetch_interval"
-    />
+    <Slider onValueChange={v => (fetchInterval = v[0])} value={[fetchInterval]} min={5} max={60} id="fetch_interval" />
   </div>
   <div class="relative mt-4 flex items-center justify-between">
     <Badge variant="outline" class="opacity-40">{app.name}@{app.version}</Badge>
-    <Button type="button" size="sm" on:click={onSave}>Save</Button>
+    <Button type="button" size="sm" onclick={onSave}>Save</Button>
   </div>
 </div>
