@@ -1,7 +1,7 @@
 <script lang="ts">
   import { Label } from '$lib/components/ui/label';
   import { appearance } from '../lib/theme';
-  let dark = $appearance.theme === 'dark';
+  let dark = $state($appearance.theme === 'dark');
   appearance.subscribe(({ theme }) => {
     dark = theme === 'dark';
   });
@@ -12,20 +12,20 @@
   <button
     id="dark-mode-toggle"
     class="{dark ? 'bg-input ' : 'bg-primary'} 
-    focus-visible:ring-ring focus-visible:ring-offset-background peer relative inline-flex flex-shrink-0 h-[24px] w-[44px] border-2 rounded-full border-transparent disabled:cursor-not-allowed focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:opacity-50 cursor-pointer transition duration-200 ease-in-out"
+    peer relative inline-flex h-[24px] w-[44px] flex-shrink-0 cursor-pointer rounded-full border-2 border-transparent transition duration-200 ease-in-out focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-background disabled:cursor-not-allowed disabled:opacity-50"
     type="button"
-    on:click={$appearance.toggle}
+    onclick={$appearance.toggle}
   >
     <span class="sr-only">Toggle Dark Mode</span>
     <span
       class="{dark
         ? 'translate-x-0 bg-background'
-        : 'translate-x-5 bg-white'} pointer-events-none relative inline-block h-5 w-5 rounded-full shadow transform ring-0 transition ease-in-out duration-200"
+        : 'translate-x-5 bg-white'} pointer-events-none relative inline-block h-5 w-5 transform rounded-full shadow ring-0 transition duration-200 ease-in-out"
     >
       <span
         class="{dark
-          ? 'opacity-100 ease-in duration-200'
-          : 'opacity-0 ease-out duration-100'} absolute inset-0 h-full w-full flex items-center justify-center transition-opacity"
+          ? 'opacity-100 duration-200 ease-in'
+          : 'opacity-0 duration-100 ease-out'} absolute inset-0 flex h-full w-full items-center justify-center transition-opacity"
         aria-hidden="true"
       >
         <!-- moon icon -->
@@ -35,8 +35,8 @@
       </span>
       <span
         class="{dark
-          ? 'opacity-0 ease-out duration-100'
-          : 'opacity-100 ease-in duration-200'} absolute inset-0 h-full w-full flex items-center justify-center transition-opacity"
+          ? 'opacity-0 duration-100 ease-out'
+          : 'opacity-100 duration-200 ease-in'} absolute inset-0 flex h-full w-full items-center justify-center transition-opacity"
         aria-hidden="true"
       >
         <!-- sun icon -->
