@@ -1,7 +1,7 @@
-import Root from "./label.svelte";
+import Root from './label.svelte';
 
 export {
-	Root,
-	//
-	Root as Label,
+  Root,
+  //
+  Root as Label,
 };
