@@ -1,11 +1,9 @@
 <script lang="ts">
-	import { Select as SelectPrimitive } from "bits-ui";
-	import { cn } from "$lib/utils.js";
+  import type { Separator as SeparatorPrimitive } from 'bits-ui';
+  import { Separator } from '$lib/components/ui/separator/index.js';
+  import { cn } from '$lib/utils.js';
 
-	type $$Props = SelectPrimitive.SeparatorProps;
-
-	let className: $$Props["class"] = undefined;
-	export { className as class };
+  let { ref = $bindable(null), class: className, ...restProps }: SeparatorPrimitive.RootProps = $props();
 </script>
 
-<SelectPrimitive.Separator class={cn("bg-muted -mx-1 my-1 h-px", className)} {...$$restProps} />
+<Separator bind:ref class={cn('-mx-1 my-1 h-px bg-muted', className)} {...restProps} />
