@@ -1,7 +1,7 @@
-import Root from "./skeleton.svelte";
+import Root from './skeleton.svelte';
 
 export {
-	Root,
-	//
-	Root as Skeleton,
+  Root,
+  //
+  Root as Skeleton,
 };
