@@ -1,7 +1,7 @@
-import Root from "./switch.svelte";
+import Root from './switch.svelte';
 
 export {
-	Root,
-	//
-	Root as Switch,
+  Root,
+  //
+  Root as Switch,
 };
