@@ -1,15 +1,17 @@
-import { Tooltip as TooltipPrimitive } from "bits-ui";
-import Content from "./tooltip-content.svelte";
+import { Tooltip as TooltipPrimitive } from 'bits-ui';
+import Content from './tooltip-content.svelte';
 
 const Root = TooltipPrimitive.Root;
+const Provider = TooltipPrimitive.Provider;
 const Trigger = TooltipPrimitive.Trigger;
 
 export {
-	Root,
-	Trigger,
-	Content,
-	//
-	Root as Tooltip,
-	Content as TooltipContent,
-	Trigger as TooltipTrigger,
+  Root,
+  Trigger,
+  Content,
+  //
+  Root as Tooltip,
+  Content as TooltipContent,
+  Trigger as TooltipTrigger,
+  Provider as TooltipProvider,
 };
