@@ -1,4 +1,4 @@
-import path from 'path';
+import { fileURLToPath } from 'url';
 import { defineConfig } from 'vite';
 import { svelte } from '@sveltejs/vite-plugin-svelte';
 
@@ -10,7 +10,7 @@ export default defineConfig({
   },
   resolve: {
     alias: {
-      $lib: path.resolve('./src/lib'),
+      $lib: fileURLToPath(new URL('./src/lib', import.meta.url)),
     },
   },
 });
